move tests to old test file
